Moved the house database tests to the right app...
